create a template folder, move standard page and layout into it
