+++ src/PageBuilderApp.tsx
@@ -1,12 +1,12 @@
 import PageRouter, { RouteData } from './PageRouter';
-import StandardLayout from './pages/StandardLayout';
+import StandardLayout from './templates/StandardLayout';
 import { AnswersHeadlessProvider } from '@yext/answers-headless-react';
 import { PageViewContextProvider } from './context/PageViewContext';
 import UniversalSearchPage from './pages/UniversalSearchPage';
 import AnswersAppConfig from './config/AnswersAppConfig';
 import { CardNameToComponentMapping } from './models/cardComponent';
 import { UniversalResultsConfig } from './config/universalResultsConfig';
-import StandardPage from './pages/StandardPage';
+import StandardPage from './templates/StandardPage';
 import { AnswersAppContextProvider } from './context/AnswersAppContext';
 
 export default function PageBuilderApp({ config }: { config: AnswersAppConfig }) {
